+++ CombatLevelDamageScaler/Mod.cs
@@ -2,7 +2,6 @@ using CombatLevelDamageScaler.Framework;
 using CombatLevelDamageScaler.Patches;
 using Spacechase.Shared.Patching;
 using SpaceShared;
-using SpaceShared.APIs;
 using StardewModdingAPI;
 using StardewModdingAPI.Events;
 
@@ -27,7 +26,7 @@ namespace CombatLevelDamageScaler
 
         private void OnGameLaunched(object sender, GameLaunchedEventArgs e)
         {
-            var configMenu = this.Helper.ModRegistry.GetApi<IGenericModConfigMenuApi>("spacechase0.GenericModConfigMenu");
+            var configMenu = this.Helper.ModRegistry.GetGenericModConfigMenuApi(this.Monitor);
             if (configMenu != null)
             {
                 configMenu.Register(