+++ ExtendedReach/Mod.cs
@@ -2,7 +2,6 @@ using ExtendedReach.Framework;
 using ExtendedReach.Patches;
 using Spacechase.Shared.Patching;
 using SpaceShared;
-using SpaceShared.APIs;
 using StardewModdingAPI;
 using StardewModdingAPI.Events;
 
@@ -45,7 +44,7 @@ namespace ExtendedReach
         *********/
         private void OnGameLaunched(object sender, GameLaunchedEventArgs e)
         {
-            var configMenu = this.Helper.ModRegistry.GetApi<IGenericModConfigMenuApi>("spacechase0.GenericModConfigMenu");
+            var configMenu = this.Helper.ModRegistry.GetGenericModConfigMenuApi(this.Monitor);
             if (configMenu != null)
             {
                 configMenu.Register(