+++ Magic/Mod.cs
@@ -77,7 +77,7 @@ namespace Magic
         {
             // hook Generic Mod Config Menu
             {
-                var configMenu = this.Helper.ModRegistry.GetApi<IGenericModConfigMenuApi>("spacechase0.GenericModConfigMenu");
+                var configMenu = this.Helper.ModRegistry.GetGenericModConfigMenuApi(this.Monitor);
                 if (configMenu != null)
                 {
                     configMenu.Register(