+++ SpaceCore/docs/release-notes.md
@@ -1,6 +1,9 @@
 ﻿﻿[← back to readme](README.md)
 
 # Release notes
+## Upcoming release
+* Added min-version check for integration with Generic Mod Config Menu.
+
 ## 1.7.2
 Released 24 December 2021 for SMAPI 3.13.0 or later. Updated by Pathoschild.